goyo-server/backend_not_in_use/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
redis==5.0.1
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
bcrypt==4.1.2
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
paho-mqtt==1.6.1
python-dotenv==1.0.0
alembic==1.13.1
pyaudio==0.2.14
numpy==1.26.3
email-validator==2.0.0
